Initial commit
I didn't need to use margin left and right: auto because justify-
content: space-between works to do just that. It spaces each box
out so I need to make the outside links into flexboxes as well
for it to push each box out and keep the inside box in the center.