Adjust boxes using shrink, basis, padding, margin
Adjusted the padding around the modal, gave spacing between items
using padding and margins, and adjusted flex properties to allow
certain containers to grow from a set basis, otherwise the icon and
close-button stretch out to fill the space. I set the header to have
a basis of auto to push the close-button out to the right.