Use containers in HTML to rearrange modal layout
Using containers in a creative way in the html is what will
set you up for success in the stylesheet. Keep the icon outside
of the container so it can be offset from the rest of the features
and use the `flex-shrink: 0;` property on the icon to keep it from
getting smashed by the container.